refactor: rename context import path and remove unused createWager context
- Updated import paths for CreateWagerProvider and useCreateWagerContext to reflect new directory structure.
- Deleted the obsolete createWager.context.tsx file as it is no longer in use.
